requirements.txt
fastapi==0.135.1
uvicorn==0.42.0
SQLAlchemy==2.0.48
asyncpg==0.31.0
pydantic==2.12.5
pydantic-settings==2.7.0
hive-nectar==0.2.14
alembic==1.14.1
sentence-transformers>=2.7,<4
numpy>=1.24
pgvector>=0.3
langdetect>=1.0.9
psycopg2-binary>=2.9
pyjwt>=2.8

# testing & benchmarks
pytest==8.3.4
pytest-asyncio==0.25.3
httpx==0.28.1
